requirements.txt
appdirs==1.4.4
asgiref==3.3.1
click==7.1.2
distlib==0.3.1
Django==3.1.7
djangorestframework==3.12.4
filelock==3.0.12
Flask==1.1.2
gitdb==4.0.7
GitPython==3.1.15
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
Pillow==8.1.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1
six==1.15.0
smmap==4.0.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
virtualenv==20.4.3
Werkzeug==1.0.1
whitenoise==5.2.0
